[Bug: #8472] - DICOM query with date filters are ignored on 64bit ''Schema/Configuration Changes: No'' * Fixed.

+++ Desktop/View/WinForms/DateTimeField.cs
@@ -182,7 +182,7 @@ namespace ClearCanvas.Desktop.View.WinForms
 		private struct NMHDR
 		{
 			public IntPtr hwndFrom;
-			public uint idFrom;
+			public IntPtr idFrom;
 			public uint code;
 		}
